- API: GET /banks to show the list of banks and their information

+++ src/main/java/io/github/rusyasoft/example/bank/ipoteka/repository/BankRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface BankRepository extends CrudRepository<Bank, Long> {
+public interface BankRepository extends CrudRepository<Bank, Integer> {
     List<Bank> findByName(String name);
 }
